Rename the yacl package to config
