+++ src/main/java/net/fellter/moreLeafParticles/config/YACLImpl.java
@@ -1,4 +1,4 @@
-package net.fellter.moreLeafParticles.yacl;
+package net.fellter.moreLeafParticles.config;
 
 import java.awt.*;
 import java.util.Arrays;
